fix compile errors for jobsrv
