+++ components/jobsrv/jobs/request/jobs.proto
@@ -1,0 +1,19 @@
+syntax = "proto3";
+
+package hurtlocker.jobsrv.jobs.request;
+option go_package = "github.com/chuckleheads/hurtlocker/components/jobsrv/jobs/request";
+
+message Job {
+    string id = 1;
+    string name = 2;
+}
+
+message GetJobsReq {};
+
+message GetJobReq {
+    string name = 1;
+}
+
+message CreateJobReq {
+    string name = 1;
+}
