--- components/jobsrv/origins/request/origins.proto
@@ -1,26 +1,0 @@
-syntax = "proto3";
-
-package hurtlocker.originsrv.origins.request;
-option go_package = "github.com/chuckleheads/hurtlocker/components/originsrv/origins/request";
-
-message Origin {
-    string id = 1;
-    string name = 2;
-    string default_package_visibility = 3;
-}
-
-message GetOriginsReq {};
-
-message GetOriginReq {
-    string name = 1;
-}
-
-message CreateOriginReq {
-    string name = 1;
-    string default_package_visibility = 2;
-}
-
-message UpdateOriginReq {
-    string name = 1;
-    string default_package_visibility = 2;
-}
